[REFACTOR] Extract/organize large RootLayout concerns in app/layout.js
